NiceController.m: fix an int-vs-NSInteger warning

+++ Classes/NiceController.m
@@ -100,7 +100,7 @@ BOOL detectIsPlaying(id each, void* context){
     return mainWindowProxy;
 }
 
--(int)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)openableFileExtensions
+-(NSInteger)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)openableFileExtensions
 {
     [openPanel setAllowsMultipleSelection:YES];
     [openPanel setCanChooseDirectories:YES];
